Not big changes
Added some function to narratives_main.py and validators.py that work with lists to display options as single letters which correspond to a full word or title (of a movie). Validators has a function now which can check to see if the user has picked a valid letter or number corresponding to their choice. I'm working on a 'complex validator' in case two list items start with the same letter.